% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkg.api-package.R
\docType{package}
\name{rpkg.api-package}
\alias{rpkg.api}
\alias{rpkg.api-package}
\title{rpkg.api: Connect to 'rpkg.net' and seek 'API'}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
